feat: 重构进程通信机制，使用WorkerPipe替代Connection，优化WebSocket和HTTP工作进程的管理

+++ server/app/processing/base.py
@@ -1,13 +1,42 @@
 from multiprocessing import Process, Queue
 import sys
-from typing import  cast
+from typing import  cast, Any
 from multiprocessing.synchronize import Event as MPEvent
 from multiprocessing import Pipe
 from multiprocessing.connection import Connection
+from dataclasses import dataclass
 
 from .log.record import WorkerLogger
 
 from .excepthook import  installProcessExcepthook, processExcepthook
+
+@dataclass
+class PipeInfo:
+    type: str
+    data: Any
+
+        
+    
+class WorkerPipe:
+    def __init__(self, pipe: Connection):
+        self.pipe = pipe
+
+        self.pipe.readable
+
+    def send(self, type: str, data: Any = None):
+        self.pipe.send(PipeInfo(type, data))
+
+    def poll(self):
+        return self.pipe.poll()
+
+    def recv(self) -> PipeInfo:
+        return self.pipe.recv()
+
+    
+
+    @classmethod
+    def make(cls,parent: Connection, child: Connection):
+        return cls(parent), cls(child) 
 
 
 
@@ -16,14 +45,18 @@ class ProcessWorker:
     def __init__(self, name: str, logQueue: Queue, stopEvent: MPEvent, daemon: bool = True):
         self.name = name
         self.daemon = daemon
+
+        # 父子共有的
         self.logQueue = logQueue
         self.stopEvent = stopEvent
 
         # 子进程创建的变量
         self.workerLogger = None
-        self.pipe: Connection = None #type: ignore
+        self.pipe: WorkerPipe = None #type: ignore
 
-        self.parentPipe: Connection = None #type: ignore
+        # 这是给父进程用的
+        self.parentPipe: WorkerPipe = None #type: ignore
+
 
         self._process = None
 
@@ -39,7 +72,7 @@ class ProcessWorker:
         '''子进程的主函数'''
         raise NotImplementedError
 
-    def _worker(self, pipe: Connection):
+    def _worker(self, pipe: WorkerPipe):
 
         try:
             installProcessExcepthook(self.getWorkerLogger())
@@ -52,9 +85,10 @@ class ProcessWorker:
 
     def start(self):
 
-        parent, children = Pipe()
+        parent, children = WorkerPipe.make(*Pipe())
 
         self.parentPipe = parent
+
 
         self._process = Process(target=self._worker, name=self.name, args=(children,), daemon=self.daemon)
 
@@ -66,8 +100,13 @@ class ProcessWorker:
 
 
     def join(self):
+        if self._process is None:
+            raise ValueError(f"{self.name} not started")
         self._process.join()
 
-    def forceStop(self):
+    def terminate(self):
+        if self._process is None:
+            raise ValueError(f"{self.name} not started")
+        
         self._process.terminate()
